feat(api, core, users)!: added bars sorting
1. Changed types of Integer Fields in user & core models.
2. Created and made migrations.
3. Created and added sorting by user's character to the bar's view.

+++ drink_n_talk/users/models.py
@@ -22,11 +22,11 @@ class User(AbstractUser):
         default='user_avatar/default_user.jpg',
         verbose_name='аватар'
     )
-    degree = models.IntegerField(
+    degree = models.PositiveSmallIntegerField(
         choices=settings.DEGREE_CHOICES,
         verbose_name='степень алкогольности'
     )
-    character = models.IntegerField(
+    character = models.SmallIntegerField(
         choices=settings.CHARACTER_CHOICES,
         verbose_name='слушатель/говорун'
     )
